requirements.txt
sc-compression
Pillow